Move Transition to a standalone class
